requirements.txt
# List of dependencies for the project, make sure that you have these installed in your Python environment before running the scripts.

numpy
pyyaml
tensorflow
tensorflow-datasets